+++ src/protobuf-net.Core/Internal/TypeHelperT.cs
@@ -58,6 +58,7 @@ namespace ProtoBuf.Internal
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0066:Convert switch statement to expression", Justification = "Readability")]
         internal static bool CanBePacked(Type type)
         {
+            type = Nullable.GetUnderlyingType(type) ?? type;
             if (type.IsEnum) return true;
             switch (Type.GetTypeCode(type))
             {
@@ -146,6 +147,14 @@ namespace ProtoBuf.Internal
                 .GetField(nameof(StructValueChecker<int>.Instance))
                 .GetValue(null);
         }
+
+        internal static object CreateNonTrivialDefault(Type type)
+        {
+            if (type.IsValueType) return Activator.CreateInstance(Nullable.GetUnderlyingType(type) ?? type);
+            if (type == typeof(string)) return "";
+            if (type == typeof(byte[])) return Array.Empty<byte>();
+            return null;
+        }
     }
 
     internal static class TypeHelper<T>
@@ -159,9 +168,11 @@ namespace ProtoBuf.Internal
             ?? ReferenceValueChecker.Instance as IValueChecker<T>
             ?? (IValueChecker<T>)TypeHelper.GetValueTypeChecker(typeof(T));
 
-        public static readonly bool CanBePacked = !CanBeNull && TypeHelper.CanBePacked(typeof(T));
+        public static readonly bool CanBePacked = !IsReferenceType && TypeHelper.CanBePacked(typeof(T));
 
         public static readonly T Default = typeof(T) == typeof(string) ? (T)(object)"" : default;
+
+        public static readonly T NonTrivialDefault = Default ?? (T)TypeHelper.CreateNonTrivialDefault(typeof(T));
 
         // make sure we don't cast null value-types to NREs
         [MethodImpl(ProtoReader.HotPath)]