+++ src/protobuf-net.Reflection/ErrorCode.cs
@@ -46,6 +46,10 @@
         UnexpectedEOF = 38,
         InvalidString = 39,
         SyntaxErrorUnknownEntity = 40,
-        InvalidUtf8 = 41,
+        MessageKindUnknownKind = 41,
+        MessageKindNullWrapperProto3 = 42,
+        MessageKindNullWrapperSingleField = 43,
+        MessageKindNullWrapperFieldOne = 44,
+        InvalidUtf8 = 45,
     }
 }