+++ src/protobuf-net.Test/Issues/IdempotentTypeModel.cs
@@ -1,6 +1,4 @@
-﻿#if FEAT_NULL_LIST_ITEMS
-
-using ProtoBuf.Meta;
+﻿using ProtoBuf.Meta;
 using System;
 using System.Collections.Generic;
 using Xunit;
@@ -13,11 +11,13 @@ namespace ProtoBuf.Issues
         public void ChangingSupportNullToSameValueWorks()
         {
             // can change any number of times before we serialize
+#pragma warning disable CS0618
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = true;
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = false;
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = true;
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = false;
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = true;
+#pragma warning restore CS0618
 
             // do the first serialization here (model becomes frozen)
             Assert.Equal("abc,def",
@@ -28,7 +28,9 @@ namespace ProtoBuf.Issues
 
             // it should not throw if we "change" it to the same value,
             // even after serialization
+#pragma warning disable CS0618
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = true;
+#pragma warning restore CS0618
 
             Assert.Equal("ghi,jkl",
                 string.Join(",", Serializer.DeepClone(
@@ -36,12 +38,16 @@ namespace ProtoBuf.Issues
                 ).List));
 
             // and again just for luck
+#pragma warning disable CS0618
             RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = true;
-            
+#pragma warning restore CS0618
+
             // but it *should* throw if a different value is used after serialization
             var ex = Assert.Throws<InvalidOperationException>(() =>
             {
+#pragma warning disable CS0618
                 RuntimeTypeModel.Default[typeof(ProtoList<string>)][1].SupportNull = false;
+#pragma warning restore CS0618
             });
             Assert.Equal(
                 "The type cannot be changed once a serializer has been generated",
@@ -58,5 +64,3 @@ namespace ProtoBuf.Issues
 
     }
 }
-
-#endif